DSA: add L2 tests and test README for DSA
For the DSA tests, we need to use custom SCC which has allow for SYS_RAWIO.
This is because the DSA workloads need to mmap the device memory and make I/O
requests. So the unpriviledged DSA applications need the capability enabled.
More details can be found on DSA guide: https://cdrdv2-public.intel.com/759709/353216-data-streaming-accelerator-user-guide-003.pdf
Signed-off-by: Manish Regmi <manish.regmi@intel.com>